backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
motor>=3.6.0
pymongo>=4.9.0
pydantic>=2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0
apscheduler==3.10.4
slowapi>=0.1.9
